+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Beaker.experiment.create()` now accepts either an `ExperimentSpec` instance or the path to a YAML spec file.
 
+### Removed
+
+- Removed `Execution` data model, since it's deprecated and redundant with `Job`.
+
 ## [v0.4.2](https://github.com/allenai/beaker-py/releases/tag/v0.4.2) - 2022-04-04
 
 ### Added